ci: split tests
ci: split tests

chore: ignore

test: make customer wallet tests serial
