+++ specs/e2e/customer-addresses-no-wallet.spec.ts
@@ -1,0 +1,180 @@
+import { faker } from "@faker-js/faker"
+
+import { test, expect } from "../fixtures/tokenizedPage"
+import { euAddress, euAddress2, euAddress3 } from "../utils/addresses"
+
+test.describe("without addresses", () => {
+  const customerEmail = faker.internet.email().toLocaleLowerCase()
+  const customerPassword = faker.internet.password()
+
+  test.use({
+    defaultParams: {
+      order: "with-items",
+      customer: {
+        email: customerEmail,
+        password: customerPassword,
+      },
+      lineItemsAttributes: [
+        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
+      ],
+    },
+  })
+
+  test("multiple address changes", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.page.locator(`text=${customerEmail}`)
+
+    await checkoutPage.setBillingAddress(euAddress)
+
+    await checkoutPage.checkStep("Customer", "open")
+
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.checkStep("Shipping", "open")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkBillingAddress(euAddress)
+
+    await checkoutPage.shipToDifferentAddress()
+
+    await checkoutPage.checkShipToDifferentAddressValue(true)
+
+    await checkoutPage.setShippingAddress()
+
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.clickStep("Customer")
+    await checkoutPage.checkBillingAddress(euAddress)
+    await checkoutPage.checkShippingAddress(euAddress2)
+
+    await checkoutPage.setBillingAddress(euAddress3)
+
+    await checkoutPage.save("Customer")
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkBillingAddress(euAddress3)
+    await checkoutPage.checkShippingAddress(euAddress2)
+
+    await checkoutPage.shipToDifferentAddress()
+
+    await checkoutPage.checkShipToDifferentAddressValue(false)
+
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkShipToDifferentAddressValue(false)
+
+    const element = await checkoutPage.page.locator(
+      'h3:has-text("Shipping Address")'
+    )
+
+    await expect(element).not.toBeVisible()
+  })
+})
+
+test.describe("same addresses", () => {
+  const customerEmail = faker.internet.email().toLocaleLowerCase()
+  const customerPassword = faker.internet.password()
+
+  test.use({
+    defaultParams: {
+      order: "with-items",
+      customer: {
+        email: customerEmail,
+        password: customerPassword,
+      },
+      lineItemsAttributes: [
+        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
+      ],
+      addresses: {
+        billingAddress: euAddress,
+        sameShippingAddress: true,
+      },
+    },
+  })
+
+  test("multiple updates", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.page.locator(`text=${customerEmail}`)
+
+    await checkoutPage.checkStep("Customer", "close")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkBillingAddress(euAddress)
+
+    await checkoutPage.setBillingAddress(euAddress3)
+
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkBillingAddress(euAddress3)
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.shipToDifferentAddress()
+
+    await checkoutPage.setShippingAddress(euAddress2)
+
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkShippingAddress(euAddress2)
+
+    await checkoutPage.shipToDifferentAddress()
+
+    await checkoutPage.checkShipToDifferentAddressValue(false)
+
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkShipToDifferentAddressValue(false)
+
+    const element = await checkoutPage.page.locator(
+      'h3:has-text("Shipping Address")'
+    )
+
+    await expect(element).not.toBeVisible()
+  })
+})
+
+test.describe("different addresses", () => {
+  const customerEmail = faker.internet.email().toLocaleLowerCase()
+  const customerPassword = faker.internet.password()
+
+  test.use({
+    defaultParams: {
+      order: "with-items",
+      customer: {
+        email: customerEmail,
+        password: customerPassword,
+      },
+      lineItemsAttributes: [
+        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
+      ],
+      addresses: {
+        billingAddress: euAddress,
+        shippingAddress: euAddress2,
+      },
+    },
+  })
+
+  test("check addresses", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.page.locator(`text=${customerEmail}`)
+
+    await checkoutPage.checkStep("Customer", "close")
+
+    await checkoutPage.clickStep("Customer")
+
+    await checkoutPage.checkBillingAddress(euAddress)
+    await checkoutPage.checkShippingAddress(euAddress2)
+  })
+})
