+++ specs/e2e/payment-sources-paypal.spec.ts
@@ -137,54 +137,3 @@ test.describe("payment source amount mismatch with paypal", () => {
     await checkoutPage.save("Payment", "Paga con PayPal")
   })
 })
-
-test.describe("payment source amount mismatch for coupon with paypal", () => {
-  const customerEmail = faker.internet.email().toLocaleLowerCase()
-
-  test.use({
-    defaultParams: {
-      order: "with-items",
-      orderAttributes: {
-        customer_email: customerEmail,
-      },
-      lineItemsAttributes: [
-        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
-      ],
-      addresses: {
-        billingAddress: euAddress,
-        sameShippingAddress: true,
-      },
-      couponCode: "testcoupon",
-    },
-  })
-
-  test("checkout remove coupon", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-
-    await checkoutPage.checkStep("Shipping", "open")
-
-    await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
-
-    await checkoutPage.page.waitForTimeout(TIMEOUT)
-
-    await checkoutPage.save("Shipping")
-    await checkoutPage.selectPayment("paypal")
-
-    await checkoutPage.page.waitForTimeout(TIMEOUT)
-
-    await checkoutPage.removeCoupon()
-    await checkoutPage.page.waitForTimeout(TIMEOUT)
-
-    await checkoutPage.checkStep("Shipping", "open")
-
-    await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
-
-    await checkoutPage.page.waitForTimeout(TIMEOUT)
-
-    await checkoutPage.save("Shipping")
-    await checkoutPage.setPayment("paypal")
-    await checkoutPage.page.waitForTimeout(TIMEOUT)
-
-    await checkoutPage.save("Payment", "Paga con PayPal")
-  })
-})
