+++ specs/e2e/shipments-errors.spec.ts
@@ -155,3 +155,67 @@ test.describe("no shipping zone and one out of stock", () => {
     await expect(element).toHaveCount(1)
   })
 })
+
+test.describe("with single shipping method, not shippable", () => {
+  test.use({
+    defaultParams: {
+      order: "with-items",
+      market: process.env.E2E_MARKET_ID_SINGLE_SHIPPING_METHOD,
+      lineItemsAttributes: [
+        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
+        { sku_code: "SOCKXXMUE63E74FFFFFFLXXX", quantity: 1 },
+      ],
+      orderAttributes: {
+        customer_email: customerEmail,
+      },
+    },
+  })
+
+  test("a shipment without available shipping methods", async ({
+    checkoutPage,
+  }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.setBillingAddress(usAddress)
+    await checkoutPage.save("Customer")
+
+    await checkoutPage.checkStep("Customer", "close")
+    await checkoutPage.checkStep("Shipping", "open")
+    await checkoutPage.checkButton({ type: "Shipping", status: "not_present" })
+    const element = checkoutPage.page.locator(
+      "text=The entered destination is outside our shipping zone."
+    )
+    await expect(element).toHaveCount(1)
+  })
+})
+
+test.describe("no shipping zone", () => {
+  test.use({
+    defaultParams: {
+      order: "with-items",
+      lineItemsAttributes: [
+        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
+      ],
+      orderAttributes: {
+        customer_email: customerEmail,
+      },
+      addresses: {
+        billingAddress: {
+          ...usAddress,
+          billing_info: faker.random.alphaNumeric(11),
+        },
+        sameShippingAddress: true,
+      },
+    },
+  })
+
+  test("no shipping method to select", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+    await checkoutPage.checkStep("Shipping", "open")
+    await checkoutPage.checkButton({ type: "Shipping", status: "not_present" })
+    const element = checkoutPage.page.locator(
+      "text=The entered destination is outside our shipping zone. Please change your shipping address or contact us for help"
+    )
+    await expect(element).toHaveCount(1)
+  })
+})
